+++ internal/pkg/disk_io/device_linux.go
@@ -3,7 +3,7 @@
 
 //go:build linux
 
-package diskio
+package disk_io
 
 import (
 	"bufio"
